Feat: Modularized most functions
